chore: extract customRender utility
